+++ client/views/components/game_list_row.rb
