+++ client/views/components/stats_list_row.rb
