+++ client/views/windows/online_game_menu_window.rb
@@ -2,7 +2,7 @@ require_relative '../../../client/views/events/window_change_event'
 require_relative '../../../client/views/observable'
 require_relative '../../../client/views/windows/main_menu_window'
 require_relative 'online_game_window'
-require_relative 'game_list_row'
+require_relative '../components/game_list_row'
 require_relative '../events/ui_event'
 require_relative 'widget_window'
 require_relative 'stats_window'
